// src/app/demo/models/modelsgeneral/nivelesviewmodel.ts
export class Niveles{
    codigo?: number;
    nive_Id?: number;
    nive_Descripcion?: string;
    usua_Creacion?: number;
    usua_Modificacion?: number;
    usuarioCreacion?: string;
    usuarioModifica?: string;
    nive_FechaCreacion?: string;
    nive_FechaModificacion?: string;

    usuaCreacion?: string;
    usuaModificacion?: string;


}